src/main/java/com/ram/jh/inventory/model/Products.java
package com.ram.jh.inventory.model;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import com.fasterxml.jackson.annotation.JsonProperty;
import java.util.List;

@JsonIgnoreProperties(ignoreUnknown = true)
public class Products {


    private List<Product> products;

    public Products(@JsonProperty("products") List<Product> products) {
        this.products = products;
    }

    public void setProducts(List<Product> products) {
        this.products = products;
    }

    public List<Product> getProducts() {
        return products;
    }

    @Override public String toString() {
        return "Products{" +
               "products=" + products +
               '}';
    }
}
